fix bio influencer controller and add authorization

+++ app/Http/Controllers/InfluencerBioController.php
@@ -2,10 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Http\Requests\InfluencerBioRequest;
 use App\Models\Category;
-use App\Models\CategoryInfluencer;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Province;
 use App\Models\Influencer;
 use App\Models\User;
@@ -16,10 +15,24 @@ class InfluencerBioController extends Controller
 {
     public function index()
     {   
+        
+       
         $provinces = Province::get(['id', 'name']);
         $logged_in_user_id = auth()->user()->id;
-        $user_data = DB::table('users')->where('id', $logged_in_user_id)->select('id', 'name', 'created_at')->first();
+        $user_data = DB::table('users')->where('id', $logged_in_user_id)->select('id', 'name', 'username', 'created_at')->first();
         $influencer = Influencer::where('user_id', $logged_in_user_id)->first();
+
+        if($influencer == null){
+            $categories = Category::get(['id', 'name']);
+            $data = [
+                'provinces' => $provinces,
+                'influencer' => $influencer,
+                'user_data' => $user_data,
+                'categories' => $categories
+            ];
+            return response()->json($data);
+        }
+
         $categories = DB::table('categories')
                     ->join('category_influencer', 'categories.id', '=', 'category_influencer.category_id')
                     ->where('category_influencer.influencer_id', $influencer->id)
@@ -91,6 +104,10 @@ class InfluencerBioController extends Controller
         $categories = $request->input('categories');
 
         $influencer = Influencer::find($id);
+
+        if(! Gate::allows('influencer_authorization', $influencer)){
+            return response()->json(['message' => 'You are not owner of this influencer data. Can be updated only by the owner.'], 403);
+        }
         $user = User::find($influencer->user_id);
 
         $profile_picture_path = $influencer->profile_picture_path;
@@ -141,7 +158,13 @@ class InfluencerBioController extends Controller
 
     public function  destroy(string $id)
     {
+        
         $influencer = Influencer::find($id);
+
+        if(! Gate::allows('influencer_authorization', $influencer)){
+            return response()->json(['message' => 'You are not owner of this influencer data. Can be deleted only by the owner.'], 403);
+        }
+
         $influencer->categories()->detach();
         $profile_picture_path = $influencer->profile_picture_path;
         $old_file_name = basename($profile_picture_path);
